i3/.config/i3/scripts/requirements.txt
pillow-simd
pyscreenshot
